requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
blinker==1.9.0
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
dataclasses-json==0.6.7
docstring_parser==0.16
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fastapi==0.115.12
filetype==1.2.0
fonttools==4.58.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
google-adk==1.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.96.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.19.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-google-genai==2.0.10
langchain-google-vertexai==2.0.24
langchain-text-splitters==0.3.8
langsmith==0.3.45
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mcp==1.9.3
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.42.1
numexpr==2.11.0
numpy==2.2.6
openpyxl==3.1.5
opentelemetry-api==1.34.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
pip==24.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pulp
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.9.1
pydeck==0.9.1
pyparsing==3.2.3
pypdf==5.6.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
validators==0.35.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
